Allow path-intermediates to be lists
This is necessary for allowing to pass arguments on to the objects at
the leafs of the data tree, which allow more than strings for item
access.
It is no issue to allow this, as dantro worked with lists internally,
anyway, and it is even better to pass on lists as it does not require
the back and forward string splitting and joining that was done up until
now.

Also: added a new test that checks if item access via lists works