Rename *_property to *_attribute; tweaks to NetworkGroup
The renaming was done to concur with the NetworkX interface names.

Also, the set_node_attributes method now allows the `from_containers`
parameter to supply containers to use instead of the members of the
network group.
Additionally, some safety measures were added, e.g. when adding
attributes from containers that do not match in length.