Implement plot creator auto detection into PlotManager
With this feature, it is possible to no longer need to specify the
`creator` key in the configuration of a plot. Instead, the
AbstractPlotCreator now has an abstract can_plot method, which can be
overwritten by derived classes. Given the internal name of the creator
and the plot configuration, it needs to evaluate whether it can create
the plot or not.
The PlotManager aggregates all the candidate creators; if it can be
unambiguously decided, which creator to use, it uses that one.

Alongside, the AbstractPlotCreator interface was changed to make those
functions used by the PlotManager public (some were private for some
reason...).

For the ExternalPlotCreator and the derived UniversePlotCreator and
MultiversePlotCreator, three factors play a role:
  - Whether the plotting function can be resolved
  - If the plott function has an attribute specifying the name of the
  creator to be used
  - If the signature of the plot function is valid, i.e. has the
  expected number and kind of arguments. This method is implemented in
  some generality and can be configured via the _AD*-prefixed class
  attributes