In UniversePlotCreator, explicitly assign order to existing ParamDims
Fixes a bug where with multiple dimensions specified, the
UniversePlotCreator failed to find the correct mapping, because the
ordering in the original ParamSpace and the ones with the coordinates
used in the plot configuration did not match. The fix is based on using
the already established order in the orginal ParamSpace for the order of
coordinate dimensions.