Add tests for the XrDataContainer with ProxyMixin
Tests were added to check wether the XrDataContainer works with proxy
support now. At the same time a bug was removed by changing the
ProxyMixin: The _postprocess_proxy_support method was removed and
instead is olnly checked in the data method after resolution if this
method is provided somewhere and applied conditionally.