Fix bugs in loading data into XarrayDataContainer
The bugs were due to the way that string data is sometimes encoded when
written by HDF5, ie.: as an array of characters, which is then loaded by
h5py as an np.ndarray of dtype 'U'. This is now converted to a python
string.

Similar issues for the `start_and_step` coordinate mode were also addressed.