Check for existing data when preparing plot configuration
This alleviates issues where a _configured_ parameter space with
many dimensions is associated with a ParamSpaceGroup and the
UniversePlotCreator then assumes all that data would be available. It is
now checked whether only the default data is available and that case is
handled in the same way as a zero-dimensional ParamSpace.

Also implements the corresponding tests.