Rework UniversePlotCreator handling of sweeps; was faulty before
Instead of using the multi-plot configuration (a ParamSpace object with
dimensions from _both_ the configuration and the available Universe
dimensions), the UniversePlotCreator should work on the _active_ state
map of _only_ the associated parameter space from the multiverse to
determine the ID of the universe to plot.

Also improves logging information to make it easier to know at which
part of the iteration the PlotManager currently is.