Implement apply_operation and use it in DAG
This also contains a flag to attempt to maintain dantro types by
inspecting the input argument type and attempting to place the result in
a container of such a type.
The most elegant way would be to have dantro types maintain their types
by themselves; however, that does not seem possible right now, which is
why such a workaround is necessary. The restriction is that the first
positional argument is assumed to hold the required type; all OPERATIONS
should be written according to this assumption.