Move PathMixin capabilities into AbstractDataContainer
The rest of dantro relies so very much on the availability of the path
mixin and it's already included into BaseDataContainer and BaseDataGroup
that it makes more sense to implement it already in their base class
(the AbstractDataContainer) such that they share the capabilities.

With this commit, the _check_data and _check_name methods also get moved
to the AbstractDataContainer; it makes more sense to have them
implemented on this deep level than in the Base* classes.

Furthermore, objects can now change their name, if they currently have
no parent associated. This is achieved by implementing the name.setter
property into the AbstractDataContainer.