Rework DAG serialization and hashing; functional now
This required to externalize the serialization and hashing to other
modules. And quite some experimenting with ruamel.yaml and its ways of
representing data was needed as well; the implementation of this commit
should be a workable solution.

Notable change: Instead of implementing the Python magic __hash__
method, the objects that are to be part of the TransformationDAG
implement their hash now as a .hashstr property. This is to not get in
the way of the int-based Python hash function.

As all hashes are now deterministic, the testing framework can include
actual comparisons of the hashes. It also demonstrates that there is no
object duplication; if the hash is the same, the same object is used.