Return results as dict instead of in BaseDataGroup
There is very little benefit in returning as a dantro group, while there
are many difficulties, especially because Link and LinkContainer have to
be used to make already attached containers or groups available in the
result group.

By using a dict, the code becomes quite a bit simpler and can
conveniently just use Python references for everything. Yeah.