Cache writing works, wohooo! :)
This is in its most basic form. It needs to be tested whether the cache
is actually also read for other objects, that's currently not clear.
Also, the circumstances in which a file cache should _not_ be written
need to be greatly expanded, otherwise too much data will be duplicated.