Various improvements to docs, tests, and edge cases
This provides full coverage of the data_ops module.

Importantly, the operations database is no longer imported from
dantro.utils, in order to not suggest it should be changed manually. To
denote that, it was refactored to `_OEPRATIONS`. The
`register_operation` function should be able to take care of all needs.

It also gets rid of some no longer needed features that were implemented
in the early days of the DAG framework, specifically the
_maintain_container_type argument; it was ditched because it is no
longer useful and unstable in itself.

Other minor changes extend test coverage and tweak some behaviour.