Generalize select_base_tag to select_base
This also changes internal storage from a tag string to a DAGReference
object, which can be identified by its type.

The motivation for this generalization is that when using a reference as
the basis of selection, there is no need to add a tag. This might be
useful when the selection base is programmatically changed, which would
otherwise overcrowd the number of tags ...