Make argument buffering and retrieval in DAG possible
This creates a significant speed-up for the concatenation operation of
the MultiversePlotCreator, because otherwise the previously created
np.ndarray of DAGReference objects needs to be re-built before
concatenation is possible, which leads to deep object inspection. With
the resolved references being xr.DataArray objects, all kinds of casting
operations are attempted, which seems to take a lot of time.

The approach taken with the argument buffering is as follows:
    - The DAGObjects database now also allows custom hashes for objects
    that do not support the `hashstr` property
    - Thus, the TransformationDAG can supply a public `add_object`
    method where any object can be added.
    - Through a custom operation, the object can be retrieved.

However, this is somewhat hacky. If there are nicer approaches, they
should be used in the MultiversePlotCreator and the add_object method
should be made private again.