Rework hashstr to use __repr__ instead of YAML -> speedup O(100)
Profiling the DAG creation led to the observation that an unnecessarily
long amount of time was spent on creating the hash string using YAML
serialization. Crucially, this time was a constant overhead of about
3-4ms and was hard to improve due to it being based on the ruamel.yaml
package.

The solution was to use the python string representation magic methods,
which are built for being (more or less) deterministic representations
of the objects. For the purpose of the hashing in dantro, the main
requirement is that they are stable and do not easily change from
version to version; this assumption might even be _safer_ for
python-based __repr__ than for the ruamel.yaml package!

With the new hashstr property, hashing occurs in ~40µs, which is a
significant speed-up!

Of course, this led to hash changes, which were all adjusted
accordingly.