Resolve Transformation.dependencies recursively from arguments
It was a bug in the previous implementation that this was not done!

The new implementation makes use of the paramspace feature
recursive_collect, which goes through all iterables and performs a type
check.
Furthermore, the property now returns a set, which required that both
Transformation, DataManager, and DAGReference are python-hashable. With
the hashstr property already existing, it was no issue to just
implemented the __hash__ property as well, which calls the python hash()
on the hash string.