Add `select_path_prefix` option to DAG; tweak profiling
The `select_path_prefix` allows specifying a path that is prepended to
the `path` given in the `select` argument. This is useful in addition to
the `select_base`.

Profiling of cache writing and lookup now only happens if there actually
*is* an attempt to write or lookup the cache. This leads to more
understandable results, i.e. zeros when there is no lookup or writing.