Configure Sphinx to be nit-picky and raise on warnings
This includes a list of to-be-ignored references, which cannot be
properly linked.

Also sets the build:docs job to allow failure and include the sphinx
error log file into the artefacts.

This furthermore changes the sphinx configuration such that it
includes __init__ docstrings (additionally to the class docstrings),
which was not the case thus far.