Select against new special `select_base` tag
This makes the `select` and `select_base` behaviour more consistent and
is overall the more general approach. The main benefit is that it
allows late evaluation of the `select` entry of meta-operations, which
is more expected than the early evaluation.

Otherwise, the use of `select` would have caused trouble for more
complex plot creators which change the select_base as part of the
procedure of adding nodes.
Furthermore, it allows registering meta-operations before the
base_transform argument is evaluated, thus allowing to use
meta-operations as part of these base transformations.