Add DataManager.dump and .restore methods
These make it possible to store a loaded data tree in a pickle file,
allowing to restore it at a later time. This way, the sometimes
time-consuming building of the data tree can be replaced by loading such
a cached file.

This required extension of some data structures, e.g. to allow
comparison of objects. With dill, the storing of the weakrefs used in
the Link object is actually not an issue, which is nice.

The tests were extended to include pickling of some built-in objects. As
no custom __setstate__ and __getstate__ methods needed to be implemented
for the majority of the data tree classes, this appears to be a
maintainable approach...