Implement ResultPlaceholder
This class is basically a `Placeholder` with a custom name. While this
could have been derived from `BaseDataProxy`, using `Placeholder` as a
base class was far easier. See #223, where consolidation is planned.

This also makes some minor changes to the proxy base class, namely that
the object argument is now no longer a requirement. This makes the
interface more flexible.