Rework __getitem__ and __contains__ for path support
This brings dantro path behaviour more closer to POSIX paths in the
following ways:

  - Can look up parent now via `..`
  - Can refer to the object itself now: `.`
  - Duplicate `/` cause no trouble

Absolute lookups still are forbidden, because it would be weird if that
would be possible.

Furthermore, `__contains__` now behaves in a more general fashion and no
longer refers only to the group itself but to the full tree beneath it.
It now answers the question: "Can this path be reached from this object?".