Add StrongLink class and use it for linked coordinates
This allows pickleability of linked coordinates at the expense of
creating more work for the garbage collector to detect the circular
references. It should still be able to do that because `StrongLink` does
not implement a custom `__del__` method; thus, the interpreter *should* be
able to safely break the circular reference.

For more info, see https://stackoverflow.com/a/10962484/1827608