Reduce deep-copying during DAG construction
The reduction is mainly during creation of the extended profile,
where the data structure requires only shallow copies. In other places,
the deep copy seems to be needed.

To allow future changes to the copying function used, there are now
module-level function names `_deepcopy` and `_shallowcopy`, which can be
changed to other implementations (e.g. pickle-based)