Move more of the DAG functionality into BasePlotCreator
The former situation was not fully consistent, this is much better now.
Also reduces the mess with passing the resolved plotting function
through as a `_plot_func` argument; instead saves it as an attribute to
the class, which makes things much easier.

Now, the plot function is resolved very early and by the base class and
the data selection and transformation is also carried out by the base
class and need not be implemented in the derived classes unless changes
are desired. The data selection also happens at a fairly late point such
that it can directly be passed to the plot function.