Remove creator-auto-detection; move plot func retrieval to PlotManager
This is another major change ... to the better. While superficially
handy to have the plot function resolved in the plot creator, it makes
much more sense to do this in the PlotManager, where other information
about the plot is aggregated as well. Most importantly: With the
auto-detection feature removed, it makes more sense to first resolve the
plot function and use that information to determine which creator to
use.