Rework and expand plot function documentation
Now many things should be documented in the `PlotManager` docs page,
because that's where the plot function is resolved. This also pertains
to some other aspects like specialisation etc.
While not fully polished, the changes made here should be a substantial
improvement to the state before, especially in combination with the new
plot function resolution.