Fix referencing issue with !dag_prev as argument for meta-operation
This only occurred in certain circumstances: When the relative reference
defined by !dag_prev was evaluated too late, i.e. after other nodes were
added.
This caused an error as soon as the argument that was specified via a
relative reference was used in a transformation *after* the first
transformation of the meta-operation.