Expand BaseDataGroup and implement file tree loaders
Implements improvements and cleanup for `BaseDataGroup`, specifically
recursive functionality that was previously available only in the
`DataManager` is now available in each group.
As a result, some methods in `DataManager` became irrelevant and were
deleted.

Furthermore, target path can now be customized in more ways, including
the relative path (and a cleaned-up version of it).

The data loaders that require class variables are now no longer
registered with the data loaders registry, because it would lead to
errors in cases where the mixin is not available.

Finally, the `FSPathLoaderMixin` is added, which makes it possible to
load filesystem trees into a flat or nested data tree structure.