Implement `UpdatePlotConfig` messaging and advanced auto-encoding
The former allows to rewrite the plots configuration from anywhere
downstream of the PlotManager, including the possibility to turn it into
a parameter sweep.

The auto-encoding-mapping is now considerably advanced, allowing Ellipses
and multi-dim-encodings. In combination with plot config rewriting, this
allows to easily define the desired `files` parameter sweep plot config.