Take into account data variables that may be used instead of dims
This makes things quite a bit more complicated for xr.Dataset-based
plotting and auto-encoding: encodings like `hue` can either be used for
dimensionality reduction via faceting -- or as an additional way to
encode a data variable, mapping it to a color.

Importantly, in `auto_encoding`, we actually need to take care how many
dimensions should be left free for the plot to handle them; free
dimensions may actually be desired!